pep8 adjust

+++ matchzoo/preprocessor/process_units.py
@@ -381,7 +381,8 @@ class FixedLengthUnit(ProcessorUnit):
         :return tokens: list of tokenized tokens in fixed length.
         """
         np_tokens = np.array(tokens)
-        fixed_tokens = np.full([self._text_length], self._pad_value)
+        fixed_tokens = np.full([self._text_length], self._pad_value,
+                               dtype=np_tokens.dtype)
 
         if self._truncate_mode == 'pre':
             trunc_tokens = tokens[-self._text_length:]
